src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-content-crash-expected.txt

CONSOLE ERROR: line 13: Uncaught TypeError: div.createShadowRoot is not a function
